backend-ai-service/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
openai>=1.40
httpx>=0.27
pydantic>=2.7
pydantic-settings>=2.3
python-dotenv>=1.0
mysql-connector-python>=9.0
